requirements2.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
ipywidgets==7.6.5
jupyter==1.0.0
jupyter_client==7.4.9
jupyter_console==6.4.3
jupyter_core==4.11.2
jupyter_server==1.4.1
jupyterlab==3.3.2
jupyterlab_pygments==0.1.2
jupyterlab_server==2.10.3
jupyterlab_widgets==1.0.0
librosa==0.7.1
numpy==1.17.3
pandas==1.3.5
scipy==1.7.3
torch==1.13.1
tqdm==4.64.1
widgetsnbextension==3.5.2
pyyaml
numba==0.48
cython
